- added NativeExecutive library for multi-threading support in rwlib - started to work on thread-safety in rwlib  * event system is not thread-safe  * windowing system is now thread-safe  * rw::Interface is not thread-safe  * DynamicTypeSystem is now thread-safe - started to work on x64 support (rwlib compiles, dependencies not yet)

+++ rwlib/src/pixelformat.hxx
@@ -987,9 +987,9 @@ inline bool doesPixelDataNeedConversion(
     // Then there is the possibility that the buffer has expanded, for any mipmap inside of pixelData.
     // A conversion will properly fix that.
     {
-        unsigned int numberOfMipmaps = pixelData.mipmaps.size();
+        size_t numberOfMipmaps = pixelData.mipmaps.size();
 
-        for ( unsigned int n = 0; n < numberOfMipmaps; n++ )
+        for ( size_t n = 0; n < numberOfMipmaps; n++ )
         {
             const pixelDataTraversal::mipmapResource& mipLayer = pixelData.mipmaps[ n ];
